Refactored home, about & faq pages.
